Re #2: Create table and sample data on Activate
Avoids double activation by adding an option in the wp_options that indicates the component has been installed prior.

Reads from the install.sql file to get the sql. It uses a regex match to get each chunk of sql statement.

To execute the sql, we are passing an array that contains a single sql statement to dbDelta(). This stops dbDelta from parsing sql statements by exploding ";". Normally, it will explode the string to array based on ";" which means that it will split sql statements even though the ";" is found in a content resulting in sql syntax error.